Moving everything to linkedlist.go for easier codegen
